downgrade gulp-rev-all to a working version 0.8.18
gulp-rev-all commit https://github.com/smysnk/gulp-rev-all/commit/056baae48d458391f173c0a1b2e8740f954fa36b
is suspected to be the cause for a bug that gives this kind of error when loading a single-page-app:
Uncaught SyntaxError: Failed to execute 'querySelector' on 'Document': '[ng-app.72f7a5f6]' is not a valid selector.

For now, solve this by downgrading to a working version of gulp-rev-all.